remove unnecessary base class
